+++ src/AzzyBot.Bot/Helpers/GeneralStrings.cs
@@ -61,6 +61,7 @@ public static class GeneralStrings
     public const string SongRequestNotFound = "This song does not exist.";
     public const string SongRequestOffline = "Because local file caching is disabled, I can't request infos about songs while your instance is offline.";
     public const string SongRequestQueued = "Your song request has been queued.";
+    public const string StationLogEmpty = "This station log is empty and cannot be viewed.";
     public const string StationNotFound = "This station does not exist.";
     public const string StationOffline = "This station is currently offline.";
     public const string StationUsersDisconnected = "All users have been disconnected from the station.";